requirements.txt
beautifulsoup4==4.9.3
python-dotenv==0.17.0
selenium==3.141.0
soupsieve==2.2.1
urllib3==1.26.4
